etl_pipleline_countries/requirements.txt
requests
psycopg2-binary
schedule
logging
load
time
config